im_client/agents/open_interpreter/requirements.txt
open-interpreter==0.2.2
fastapi
uvicorn
docker
aiohttp